feat: 🎸 서버로 메세지 전송 성공
✅ Closes: #59
