+++ client/src/components/Chatting/index.tsx
@@ -1,8 +1,9 @@
 /** @jsxImportSource @emotion/react */
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import ChatLog from "./ChatLog";
 import ChatInput from "./ChatInput";
 import * as style from "./styles";
+import io from "socket.io-client";
 
 const dummyLog = [
   {
@@ -43,18 +44,57 @@ export type Chat = {
   sendTime: Date;
 };
 
+const socket = io(`${process.env.REACT_APP_SOCKET_SERVER_URL}`, {
+  // websocket으로 먼저 연결 시도 후 실패 시 polling으로 연결
+  transports: ["websocket", "polling"],
+  autoConnect: false,
+});
+
 const Chatting = () => {
   const [logs, setLogs] = useState<Chat[]>(dummyLog);
+  const [message, setMessage] = useState<string>("");
+
+  const sendMessage = () => {
+    if (message === "") {
+      alert("메세지를 입력하세요");
+      return;
+    }
+    const newLog: Chat = {
+      sender: "me",
+      message,
+      sendTime: new Date(),
+    };
+
+    socket.emit("chat/lobby", newLog);
+    setMessage("");
+  };
+
+  const pressEnter = (e: React.KeyboardEvent<HTMLInputElement>) => {
+    if (e.key === "Enter") {
+      sendMessage();
+    }
+  };
 
   const addLogs = (newLog: Chat) => {
     setLogs((current: Chat[]) => [...current, newLog]);
+    console.log(logs);
   };
+
+  useEffect(() => {
+    socket.connect();
+  }, []);
 
   return (
     <div css={style.ChattingContainerStyle}>
       <ChatLog logs={logs} />
       <hr css={style.ChattingHRStyle} />
-      <ChatInput addLogs={addLogs} />
+      <ChatInput
+        addLogs={addLogs}
+        sendMessage={sendMessage}
+        pressEnter={pressEnter}
+        message={message}
+        setMessage={setMessage}
+      />
     </div>
   );
 };
